requirements.txt
streamlit
requests
beautifulsoup4
rank-bm25
sentence-transformers
langchain-chroma
langchain-huggingface
chromadb
PyMuPDF
python-docx
urllib3